Move rename test
